dom/indexedDB/test/test_filehandle_lifetimes_nested.html

<!--
  Any copyright is dedicated to the Public Domain.
  http://creativecommons.org/publicdomain/zero/1.0/
-->
<html>
<head>
  <title>File Handle Test</title>

  <script type="text/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>

  <script type="text/javascript;version=1.7">
  function testSteps()
  {
    const name = window.location.pathname;

    let request = indexedDB.open(name, 1);
    request.onerror = errorHandler;
    request.onsuccess = grabEventAndContinueHandler;
    let event = yield undefined;

    let db = event.target.result;
    db.onerror = errorHandler;

    request = db.createMutableFile("test.txt");
    request.onerror = errorHandler;
    request.onsuccess = grabEventAndContinueHandler;
    event = yield undefined;

    let mutableFile = event.target.result;
    mutableFile.onerror = errorHandler;

    let fileHandle = mutableFile.open();

    let fileHandle2;

    let comp = SpecialPowers.wrap(SpecialPowers.Components);
    let thread = comp.classes["@mozilla.org/thread-manager;1"]
                     .getService(comp.interfaces.nsIThreadManager)
                     .currentThread;

    let eventHasRun;

    thread.dispatch(function() {
      eventHasRun = true;

      fileHandle2 = mutableFile.open();
    }, SpecialPowers.Ci.nsIThread.DISPATCH_NORMAL);

    while (!eventHasRun) {
      thread.processNextEvent(false);
    }

    ok(fileHandle2, "Non-null fileHandle2");

    continueToNextStep();
    yield undefined;

    finishTest();
    yield undefined;
  }
  </script>
  <script type="text/javascript;version=1.7" src="file.js"></script> <!-- This is included just to skip this test in test_ipc.html -->
  <script type="text/javascript;version=1.7" src="helpers.js"></script>

</head>

<body onload="runTest();"></body>

</html>
